requirements.txt
webdataset
pandas
tqdm
pillow
torch
torchvision